requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pillow==10.3.0
sentence-transformers==3.0.1
hnswlib==0.8.0
numpy==1.26.4
python-multipart==0.0.9
orjson==3.10.7
tqdm==4.66.1
apscheduler==3.10.4
opencv-python-headless==4.10.0.84
